requirements.txt
Django>=3.0
djangorestframework
psycopg2-binary
djangorestframework-simplejwt
